fix: `bugprone-narrowing-conversions` surrounding `tr_session_stats::uploadedBytes`

+++ qt/StatsDialog.cc
@@ -53,12 +53,12 @@ void StatsDialog::updateStats()
     ui_.currentUploadedValueLabel->setText(Formatter::storage_to_string(current.uploadedBytes));
     ui_.currentDownloadedValueLabel->setText(Formatter::storage_to_string(current.downloadedBytes));
     ui_.currentRatioValueLabel->setText(Formatter::ratio_to_string(current.ratio));
-    ui_.currentDurationValueLabel->setText(Formatter::time_to_string(current.secondsActive));
+    ui_.currentDurationValueLabel->setText(Formatter::time_to_string(static_cast<int>(current.secondsActive)));
 
     ui_.totalUploadedValueLabel->setText(Formatter::storage_to_string(total.uploadedBytes));
     ui_.totalDownloadedValueLabel->setText(Formatter::storage_to_string(total.downloadedBytes));
     ui_.totalRatioValueLabel->setText(Formatter::ratio_to_string(total.ratio));
-    ui_.totalDurationValueLabel->setText(Formatter::time_to_string(total.secondsActive));
+    ui_.totalDurationValueLabel->setText(Formatter::time_to_string(static_cast<int>(total.secondsActive)));
 
     ui_.startCountLabel->setText(tr("Started %Ln time(s)", nullptr, total.sessionCount));
 }
